change #B9 add dish is working home changes

+++ restaurantmysql/src/main/java/il/co/electriccollege/restaurant/yossi/sql/dao/Breakfast.java
@@ -4,7 +4,7 @@ public class Breakfast extends AbstractDish{
 
     String eggs;
 
-    public Breakfast(String name, String description, float price, Category category, String eggs) {
+    public Breakfast(String name, String description, double price, Category category, String eggs) {
         super(name, description, price, category);
         this.eggs = eggs;
     }
